+++ src/types/geometrycollection.rs
@@ -106,7 +106,7 @@ where
             _ => return Err("Expected a word in GEOMETRYCOLLECTION"),
         };
 
-        let item = Wkt::from_word_and_tokens(&word, tokens)?;
+        let item = Wkt::from_word_and_tokens(word, tokens)?;
         items.push(item);
 
         while let Some(&Ok(Token::Comma)) = tokens.peek() {
@@ -117,7 +117,7 @@ where
                 _ => return Err("Expected a word in GEOMETRYCOLLECTION"),
             };
 
-            let item = Wkt::from_word_and_tokens(&word, tokens)?;
+            let item = Wkt::from_word_and_tokens(word, tokens)?;
             items.push(item);
         }